railroads: Check in vappend-block-layout
Layouts need to be able to compute tip ys from specs, even though they
are initialized with concrete tips. (It doesn't make sense for the
diagram layer to do the computation in the tip-y augmentation in this
commit.) Layouts need to know the number of logical rows for this.

Decided that physical rows don't make sense due to continuous
tips. Physical tip specs will be [0, 1].